Added a js/css animation "framework"
All that is needed for this to work is that one adds two css classes:
  animate AND a animation-class.

data-delay and data-duration is optional.
All the animations are done with css, the javascript just toggles them
to be triggered when they are in the viewport.

All transitions are borrowed from [http://daneden.me/animate](http://daneden.me/animate)