Partial updates and about us
Added a temporary about-us page. Made remote assistance useable.
The top-image partial has gotten more customisable, and it's added
instead of rewriting the HTML it's using (on coworker, coworkers and so
on).

The links are not in the iteam-yellow color anymore, since it's not
following the visal identiy guidelines. Links are now underlined, and
when active they should have a green border underneath it.

What's left to do?
There should be a proper About Us.
The site SHOULD be working on > IE9 (or there should be a temporary page
leading to remote assi...).
Mobile menu should work as expected.